# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.httpError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  gratitude.GetAllInboxResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/gratitude.Message'
        type: array
    type: object
  gratitude.GetAllOutboxResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/gratitude.Message'
        type: array
    type: object
  gratitude.GetAllPublicResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/gratitude.Message'
        type: array
    type: object
  gratitude.Message:
    properties:
      id:
        type: string
      kind:
        type: string
      recipientIDs:
        items:
          type: string
        type: array
      recipients:
        items:
          type: string
        type: array
      sender:
        type: string
      sender_id:
        type: string
      sent_at:
        type: string
      text:
        type: string
    type: object
  gratitude.SendResponse:
    properties:
      id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API to handle CRUD operations for user/public messages.
  license: {}
  title: Gratitude API
  version: "1.0"
paths:
  /health:
    get:
      consumes:
      - application/json
      description: check the servers health status
      operationId: Health
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.httpError'
      security:
      - BasicAuth: []
      summary: Server health check
  /inbox:
    get:
      consumes:
      - application/json
      description: retrieve all messages in a users inbox
      operationId: Get Inbox
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gratitude.GetAllInboxResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.httpError'
      security:
      - BasicAuth: []
      summary: Retrieve a users inbox
  /outbox:
    get:
      consumes:
      - application/json
      description: retrieve all messages in users outbox
      operationId: Get Inbox
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gratitude.GetAllOutboxResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.httpError'
      security:
      - BasicAuth: []
      summary: Retrieve a users outbox
  /private:
    post:
      consumes:
      - application/json
      description: send a user a private message
      operationId: Send Private
      parameters:
      - description: message
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/gratitude.Message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gratitude.SendResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.httpError'
      security:
      - BasicAuth: []
      summary: Send a user a private message
  /public:
    get:
      consumes:
      - application/json
      description: retrieve all public messages
      operationId: Get Public
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gratitude.GetAllPublicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.httpError'
      security:
      - BasicAuth: []
      summary: Retrieve all public messages
    post:
      consumes:
      - application/json
      description: send a public message
      operationId: Send Private
      parameters:
      - description: message
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/gratitude.Message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gratitude.SendResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.httpError'
      security:
      - BasicAuth: []
      summary: Send a public message
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"
x-extension-openapi:
  example: value on a json format